Add favorite extensions file.
https://github.com/Esri/arcgis-maps-sdk-swift-samples/pull/313#discussion_r1425693648
